8x speed up by replacing `py_function`

New
```
Examples/sec (First included) 11048.54 ex/sec (total: 200736 ex, 18.17 sec)
Examples/sec (First only) 11.25 ex/sec (total: 32 ex, 2.84 sec)
Examples/sec (First excluded) 13096.36 ex/sec (total: 200704 ex, 15.33 sec)
```

Previous
```
Examples/sec (First included) 1471.78 ex/sec (total: 200736 ex, 136.39 sec)
Examples/sec (First only) 11.37 ex/sec (total: 32 ex, 2.81 sec)
Examples/sec (First excluded) 1502.54 ex/sec (total: 200704 ex, 133.58 sec)
```

Curious why token_id should be float32... I kept it any way.